+++ pycryptobot/models/BotConfig.py
@@ -20,6 +20,7 @@ from pycryptobot.models.config import (
     dummyConfigParser,
     loggerConfigParser,
 )
+from pycryptobot.models.exchange.Granularity import Granularity
 from pycryptobot.models.helper.LogHelper import Logger
 from pycryptobot.models.exchange.ExchangesEnum import Exchange
 
@@ -28,7 +29,7 @@ class BotConfig:
     def __init__(self, app: CementApp, exchange: str):
         self.cli_args = app.pargs
         self.configbuilder = False
-        self.granularity = 3600
+        self.granularity = Granularity.ONE_HOUR
         self.base_currency = "BTC"
         self.quote_currency = "GBP"
         self.is_live = 0
@@ -47,8 +48,9 @@ class BotConfig:
         self.sell_at_loss = 1
         self.smart_switch = 1
         self.telegram = False
-        self.telegramdatafolder = ""
-        self.logbuysellinjson = False
+
+        self.telegramdatafolder = "."
+
         self.buypercent = 100
         self.sellpercent = 100
         self.last_action = None
@@ -79,10 +81,16 @@ class BotConfig:
         self.disabletracker = False
         self.enableml = False
         self.websocket = False
+        self.enableexitaftersell = False
 
         self.enableinsufficientfundslogging = False
         self.insufficientfunds = False
         self.enabletelegrambotcontrol = False
+        self.enableimmediatebuy = False
+
+        self.telegramtradesonly = False
+        self.disabletelegramerrormsgs = False
+
 
         self.filelog = True
         self.logfile = self.cli_args.logfile if 'logfile' in self.cli_args else 'pycryptobot.log'
@@ -111,6 +119,10 @@ class BotConfig:
         # set exchange platform
         self.exchange = self._set_exchange(exchange)
 
+        self.startmethod = self.cli_args["startmethod"] if "startmethod" in self.cli_args and self.cli_args["startmethod"] else "standard"
+        self.enable_atr72_pcnt = True
+        self.enable_buy_next = True
+        self.enable_volume = False
         # set defaults
         (
             self.api_url,
@@ -120,18 +132,73 @@ class BotConfig:
             self.market,
         ) = self._set_default_api_info(self.exchange)
 
+        self.read_config(self.exchange)
+
+        Logger.configure(
+            filelog=self.filelog,
+            logfile=self.logfile,
+            fileloglevel=self.fileloglevel,
+            consolelog=self.consolelog,
+            consoleloglevel=self.consoleloglevel,
+        )
+
+    # read and set config from file
+    def read_config(self, exchange):
+        if os.path.isfile(self.config_file):
+            self.config_provided = True
+            try:
+                with open(self.config_file, "r", encoding="utf8") as stream:
+                    try:
+                        self.config = yaml.safe_load(stream)
+                    except:
+                        try:
+                            stream.seek(0)
+                            self.config = json.load(stream)
+                        except json.decoder.JSONDecodeError as err:
+                            sys.tracebacklimit = 0
+                            raise ValueError(f"Invalid config.json: {str(err)}")
+
+            except (ScannerError, ConstructorError) as err:
+                sys.tracebacklimit = 0
+                raise ValueError(
+                    f"Invalid config: cannot parse config file: {str(err)}"
+                )
+
+            except (IOError, FileNotFoundError) as err:
+                sys.tracebacklimit = 0
+                raise ValueError(f"Invalid config: cannot open config file: {str(err)}")
+
+            except ValueError as err:
+                sys.tracebacklimit = 0
+                raise ValueError(f"Invalid config: {str(err)}")
+
+            except:
+                raise
+
+            # set exchange platform
+        self.exchange = self._set_exchange(exchange)
+
+            # set defaults
+        (
+            self.api_url,
+            self.api_key,
+            self.api_secret,
+            self.api_passphrase,
+            self.market,
+        ) = self._set_default_api_info(self.exchange)
+
         if self.config_provided:
-            if self.exchange == Exchange.COINBASEPRO.value and "coinbasepro" in self.config:
-                coinbaseProConfigParser(self, self.config["coinbasepro"], self.cli_args)
+            if self.exchange == Exchange.COINBASEPRO and self.exchange.value in self.config:
+                coinbaseProConfigParser(self, self.config[self.exchange.value], self.cli_args)
 
-            elif self.exchange == Exchange.BINANCE.value and "binance" in self.config:
-                binanceConfigParser(self, self.config["binance"], self.cli_args)
+            elif self.exchange == Exchange.BINANCE and self.exchange.value in self.config:
+                binanceConfigParser(self, self.config[self.exchange.value], self.cli_args)
 
-            elif self.exchange == Exchange.KUCOIN.value and "kucoin" in self.config:
-                kucoinConfigParser(self, self.config["kucoin"], self.cli_args)
+            elif self.exchange == Exchange.KUCOIN and self.exchange.value in self.config:
+                kucoinConfigParser(self, self.config[self.exchange.value], self.cli_args)
 
-            elif self.exchange == Exchange.DUMMY.value and "dummy" in self.config:
-                dummyConfigParser(self, self.config["dummy"], self.cli_args)
+            elif self.exchange == Exchange.DUMMY and self.exchange.value in self.config:
+                dummyConfigParser(self, self.config[self.exchange.value], self.cli_args)
 
             if (
                     not self.disabletelegram
@@ -141,9 +208,15 @@ class BotConfig:
             ):
                 telegram = self.config["telegram"]
                 self._chat_client = Telegram(telegram["token"], telegram["client_id"])
-                if "datafolder" in self.config["telegram"]:
+                if "datafolder" in telegram:
                     self.telegramdatafolder = telegram["datafolder"]
                 self.telegram = True
+
+            if "scanner" in self.config:
+                self.enableexitaftersell = self.config["scanner"]["enableexitaftersell"] if "enableexitaftersell" in self.config["scanner"] else False
+                self.enable_buy_next = True if "enable_buy_now" not in self.config["scanner"] else self.config["scanner"]["enable_buy_now"]
+                self.enable_atr72_pcnt = True if "enable_atr72_pcnt" not in self.config["scanner"] else self.config["scanner"]["enable_atr72_pcnt"]
+                self.enable_volume = False if "enable_volume" not in self.config["scanner"] else self.config["scanner"]["enable_volume"]
 
             if "logger" in self.config:
                 loggerConfigParser(self, self.config["logger"])
@@ -154,9 +227,9 @@ class BotConfig:
                 self.logfile == "/dev/null"
 
         else:
-            if self.exchange == Exchange.BINANCE.value:
+            if self.exchange == Exchange.BINANCE:
                 binanceConfigParser(self, None, self.cli_args)
-            elif self.exchange == Exchange.KUCOIN.value:
+            elif self.exchange == Exchange.KUCOIN:
                 kucoinConfigParser(self, None, self.cli_args)
             else:
                 coinbaseProConfigParser(self, None, self.cli_args)
@@ -165,40 +238,27 @@ class BotConfig:
             self.fileloglevel = "NOTSET"
             self.logfile == "/dev/null"
 
-        Logger.configure(
-            filelog=self.filelog,
-            logfile=self.logfile,
-            fileloglevel=self.fileloglevel,
-            consolelog=self.consolelog,
-            consoleloglevel=self.consoleloglevel,
-        )
 
-    def _set_exchange(self, exchange: str = None) -> str:
-        valid_exchanges = ["coinbasepro", "binance", "kucoin", "dummy"]
+    def _set_exchange(self, exchange: str = None) -> Exchange:
 
-        if 'exchange' in self.cli_args and self.cli_args.exchange is not None:
-            exchange = self.cli_args.exchange
+        if self.cli_args["exchange"] is not None:
+            exchange = Exchange(self.cli_args["exchange"])
 
-        if exchange and exchange in valid_exchanges:
-            return exchange
+        if isinstance(exchange, str):
+            exchange = Exchange(exchange)
 
         if not exchange:
-            if ("coinbasepro" or "api_pass") in self.config:
-                exchange = "coinbasepro"
-            elif "binance" in self.config:
-                exchange = "binance"
-            elif "kucoin" in self.config:
-                exchange = "kucoin"
+            if (Exchange.COINBASEPRO.value or "api_pass") in self.config:
+                exchange = Exchange.COINBASEPRO
+            elif Exchange.BINANCE.value in self.config:
+                exchange = Exchange.BINANCE
+            elif Exchange.KUCOIN.value in self.config:
+                exchange = Exchange.KUCOIN
             else:
-                exchange = "dummy"
-
-        if exchange not in valid_exchanges:
-            raise TypeError(
-                f"Invalid exchange: {exchange}. Valid choices: {valid_exchanges}"
-            )
+                exchange = Exchange.DUMMY
         return exchange
 
-    def _set_default_api_info(self, exchange: str = "dummy") -> tuple:
+    def _set_default_api_info(self, exchange: Exchange = Exchange.DUMMY) -> tuple:
         conf = {
             "binance": {
                 "api_url": "https://api.binance.com",
@@ -230,11 +290,11 @@ class BotConfig:
             },
         }
         return (
-            conf[exchange]["api_url"],
-            conf[exchange]["api_key"],
-            conf[exchange]["api_secret"],
-            conf[exchange]["api_passphrase"],
-            conf[exchange]["market"],
+            conf[exchange.value]["api_url"],
+            conf[exchange.value]["api_key"],
+            conf[exchange.value]["api_secret"],
+            conf[exchange.value]["api_passphrase"],
+            conf[exchange.value]["market"],
         )
 
     def getVersionFromREADME(self) -> str:
